\name{extractoR.mongo-package}
\alias{extractoR.mongo-package}
\alias{extractoR.mongo}
\docType{package}
\encoding{UTF-8}
\title{R packages metadata extractor (mongo)}
\description{
  This package is part of extractoR which is a set of R packages that
  can be used to fetch, extract and query R packages
  metadata. extractoR.mongo handle queries to MongoDB.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\examples{
\dontrun{
TODO
}
}